+++ retrofit/build.gradle
@@ -3,6 +3,7 @@ apply plugin: 'com.android.library'
 
 ext {
     projectName = "Retrofit"
+    versionCodeId = 1
     versionCodeNumber = "0.1.0"
     PUBLISH_GROUP_ID = applicationId
     PUBLISH_ARTIFACT_ID = "retrofit"
@@ -47,15 +48,15 @@ android {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile project(':core')
-    apt "org.projectlombok:lombok:${lombokVersion}"
-    compile "com.jakewharton.threetenabp:threetenabp:${localDateBackportVersion}"
-    compile "com.squareup.retrofit2:converter-gson:${retrofitVersion}"
-    compile "com.squareup.retrofit2:adapter-rxjava2:${retrofitVersion}"
-    compile "com.squareup.retrofit2:retrofit:${retrofitVersion}"
-    compile "io.reactivex.rxjava2:rxandroid:${rxAndroidVersion}"
-    compile "io.reactivex.rxjava2:rxjava:${rxJavaVersion}"
-    provided "org.projectlombok:lombok:${lombokVersion}"
-    testCompile "junit:junit:${junitVersion}"
+    apt "org.projectlombok:lombok:${libraryVersion.lombok}"
+    compile "com.jakewharton.threetenabp:threetenabp:${libraryVersion.localDateBackport}"
+    compile "com.squareup.retrofit2:converter-gson:${libraryVersion.retrofit}"
+    compile "com.squareup.retrofit2:adapter-rxjava2:${libraryVersion.retrofit}"
+    compile "com.squareup.retrofit2:retrofit:${libraryVersion.retrofit}"
+    compile "io.reactivex.rxjava2:rxandroid:${libraryVersion.rxAndroid}"
+    compile "io.reactivex.rxjava2:rxjava:${libraryVersion.rxJava}"
+    provided "org.projectlombok:lombok:${libraryVersion.lombok}"
+    testCompile "junit:junit:${libraryVersion.junit}"
     testCompile ('org.threeten:threetenbp:1.3.2') {
         exclude group:'com.jakewharton.threetenabp', module:'threetenabp'
     }
