+++ retrofit/src/main/java/com/xmartlabs/base/retrofit/module/ServiceGsonModule.java
@@ -9,9 +9,10 @@ import com.google.gson.FieldAttributes;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.xmartlabs.base.core.model.BuildInfo;
-import com.xmartlabs.base.retrofit.adapter.MillisecondsLocalDateAdapter;
-import com.xmartlabs.base.retrofit.adapter.MillisecondsLocalDateTimeAdapter;
-import com.xmartlabs.base.retrofit.common.GsonExclude;
+import com.xmartlabs.base.core.adapter.MillisecondsLocalDateAdapter;
+import com.xmartlabs.base.core.adapter.MillisecondsLocalDateTimeAdapter;
+import com.xmartlabs.base.core.common.GsonExclude;
+import com.xmartlabs.base.core.module.GsonExclusionStrategy;
 import com.xmartlabs.base.retrofit.deserialized.RequiredFieldDeserializer;
 
 import org.threeten.bp.LocalDate;
@@ -24,7 +25,7 @@ import dagger.Module;
 import dagger.Provides;
 
 @Module
-public class ServiceGsonModule {
+public class ServiceGsonModule extends GsonExclusionStrategy {
   public static final String SERVICE_GSON_NAME = "ServiceGson";
 
   @Named(SERVICE_GSON_NAME)
@@ -32,33 +33,9 @@ public class ServiceGsonModule {
   @Singleton
   @SuppressWarnings("unused")
   public Gson provideServiceGson(@NonNull GsonBuilder gsonBuilder, @NonNull BuildInfo buildInfo) {
-    modifyGsonBuilder(gsonBuilder, buildInfo)
+    return modifyGsonBuilder(gsonBuilder, buildInfo)
         .setExclusionStrategies(getExclusionStrategy(GsonExclude.Strategy.SERVICE))
-        .registerTypeAdapter(LocalDateTime.class, new MillisecondsLocalDateTimeAdapter())
-        .registerTypeAdapter(LocalDate.class, new MillisecondsLocalDateAdapter());
-
-    return gsonBuilder.create();
-  }
-
-  private ExclusionStrategy getExclusionStrategy(@Nullable GsonExclude.Strategy strategy) {
-    return new ExclusionStrategy() {
-      @Override
-      public boolean shouldSkipField(FieldAttributes fieldAttributes) {
-        return shouldSkipFieldFromSerialization(fieldAttributes)
-            || (fieldAttributes.getAnnotation(GsonExclude.class) != null
-            && (Objects.equals(fieldAttributes.getAnnotation(GsonExclude.class).strategy(), GsonExclude.Strategy.ALL)
-            || Objects.equals(fieldAttributes.getAnnotation(GsonExclude.class).strategy(), strategy)));
-      }
-
-      @Override
-      public boolean shouldSkipClass(Class<?> clazz) {
-        return false;
-      }
-    };
-  }
-
-  protected boolean shouldSkipFieldFromSerialization(@NonNull FieldAttributes fieldAttributes) {
-    return false;
+        .create();
   }
 
   protected GsonBuilder modifyGsonBuilder(@NonNull GsonBuilder builder, @NonNull BuildInfo buildInfo) {
@@ -66,7 +43,6 @@ public class ServiceGsonModule {
       builder.setPrettyPrinting();
     }
     builder.registerTypeHierarchyAdapter(Object.class, new RequiredFieldDeserializer());
-    builder.setExclusionStrategies(getExclusionStrategy(null));
     return builder;
   }
 }
