+++ versions.gradle
@@ -1,6 +1,5 @@
 ext {
     applicationId = "com.xmartlabs.base"
-    versionCodeId = 1
 
     compileSdkVersion = 25
     minSdkVersion = 17
@@ -8,32 +7,34 @@ ext {
     buildToolsVersion = '25.0.2'
     javaVersion8 = JavaVersion.VERSION_1_8
 
-    androidSupportVersion = '25.3.1'
-    androidSupportTestVersion = '0.5'
-    bulletVersion = 'e5044b53df'
-    butterKnifeVersion = '8.2.1'
-    daggerVersion = '2.6'
-    dartVersion = '2.0.0'
-    dexmakerVersion = '1.4'
-    dbFlowVersion = '3.1.1'
-    espressoVersion = '2.2.2'
-    firebaseVersion = '9.2.1'
-    fragmentArgsVersion = '3.0.2'
-    gsonVersion = '1.7.2'
-    jsr305Version = '3.0.1'
-    junitVersion = '4.12'
-    localDateBackportVersion = '1.0.4'
-    lombokVersion = '1.16.10'
-    mockitoVersion = '1.10.19'
-    okHttpVersion = '3.6.0'
-    okioVersion = '1.11.0'
-    playServicesVersion = '10.2.1'
-    parcelerVersion = '1.1.5'
-    restMockVersion = '0.1.3'
-    retrofitVersion = '2.2.0'
-    rxAndroidVersion = '2.0.1'
-    rxJavaVersion = '2.0.7'
-    rxLifecycleVersion = '2.0.1'
-    streamVersion = '1.1.6'
-    timberVersion = '4.3.0'
+    libraryVersion = [
+        androidSupport: '25.3.1',
+        androidSupportTest: '0.5',
+        bullet: 'e5044b53df',
+        butterKnife: '8.2.1',
+        dagger: '2.6',
+        dart: '2.0.0',
+        dexmaker: '1.4',
+        dbFlow: '3.1.1',
+        espresso: '2.2.2',
+        firebase: '9.2.1',
+        fragmentArgs: '3.0.2',
+        gson: '1.7.2',
+        jsr305: '3.0.1',
+        junit: '4.12',
+        localDateBackport: '1.0.4',
+        lombok: '1.16.10',
+        mockito: '1.10.19',
+        okHttp: '3.6.0',
+        okio: '1.11.0',
+        playServices: '10.2.1',
+        parceler: '1.1.5',
+        restMock: '0.1.3',
+        retrofit: '2.2.0',
+        rxAndroid: '2.0.1',
+        rxJava: '2.0.7',
+        rxLifecycle: '2.0.1',
+        stream: '1.1.6',
+        timber: '4.3.0',
+    ]
 }
